Implement chain of responsibility pattern for move validation in chess game
- Added OccupancyValidator to check if a destination position is occupied by a piece of the same color.
- Introduced PawnPromotionValidator to handle pawn promotion logic.
- Created PieceMovementValidator to validate moves according to chess piece movement rules.
- Implemented StalemateValidator to check for stalemate conditions.
- Added ThreefoldRepetitionValidator to validate the threefold repetition rule.
- Updated validators.dart to export all validators for easy access.
- Developed GameStateContext to manage game states (Waiting, Playing, Paused, Finished).
- Introduced AdaptiveAIStrategy that learns from player moves and adjusts its strategy accordingly.
- Enhanced AdvancedMinimaxAIStrategy with better evaluation metrics for AI decision-making.
- Created ChessAIPlayer to manage AI strategy and provide move hints to players.
- Implemented RandomAIStrategy for random move selection.
- Added move evaluation logic to support AI strategies.
