+++ lib/presentation/game_room/bloc/game_room_bloc.dart
@@ -6,8 +6,8 @@ import 'package:chess_game/core/models/chess_piece.dart';
 import 'package:chess_game/core/models/position.dart';
 import 'package:chess_game/core/models/game_room.dart';
 import 'package:chess_game/core/models/game_config.dart';
-import 'package:chess_game/core/patterns/strategy/ai_strategy.dart';
-import 'package:chess_game/core/patterns/chain_of_responsibility/move_validator.dart';
+import 'package:chess_game/presentation/game_room/strategy/ai_strategy.dart';
+import 'package:chess_game/presentation/game_room/chain_of_responsibility/move_validator.dart';
 import 'package:chess_game/presentation/game_room/command/move_command.dart';
 import 'package:chess_game/presentation/game_room/command/game_room_command_manager.dart';
 import 'package:chess_game/presentation/game_room/memento/chess_board_manager.dart';
