+++ lib/presentation/game_room/chain_of_responsibility/move_validator.dart
@@ -1,0 +1,3 @@
+// Re-export all validators and related classes for backward compatibility
+export 'validators/validators.dart';
+export 'game_room_move_validator.dart';
