+++ lib/presentation/game_room/chain_of_responsibility/validators/actual_move_validator.dart
@@ -1,0 +1,13 @@
+import 'package:chess_game/core/models/chess_piece.dart';
+import 'package:chess_game/core/models/position.dart';
+import 'move_validator_base.dart';
+
+/// 1. Validates the piece is actually moving (from != to)
+class ActualMoveValidator extends MoveValidator {
+  @override
+  bool handleValidation(
+      ChessPiece piece, Position from, Position to, List<ChessPiece> allPieces,
+      [FIDERuleContext? context]) {
+    return from != to;
+  }
+}
