+++ lib/presentation/game_room/chain_of_responsibility/validators/bounds_validator.dart
@@ -1,0 +1,13 @@
+import 'package:chess_game/core/models/chess_piece.dart';
+import 'package:chess_game/core/models/position.dart';
+import 'move_validator_base.dart';
+
+/// 2. Validates if the position is inside the board
+class BoundsValidator extends MoveValidator {
+  @override
+  bool handleValidation(
+      ChessPiece piece, Position from, Position to, List<ChessPiece> allPieces,
+      [FIDERuleContext? context]) {
+    return to.row >= 0 && to.row < 8 && to.col >= 0 && to.col < 8;
+  }
+}
