+++ lib/presentation/game_room/strategy/ai_strategy.dart
@@ -1,0 +1,9 @@
+// Export all AI strategy components for backward compatibility
+export 'move_evaluation.dart';
+export 'ai_strategy_interface.dart';
+export 'base_ai_strategy.dart';
+export 'chess_ai_player.dart';
+export 'random_ai_strategy.dart';
+export 'minimax_ai_strategy.dart';
+export 'advanced_minimax_ai_strategy.dart';
+export 'adaptive_ai_strategy.dart';
